requirements.txt
python-telegram-bot==21.7
python-dotenv
pyVmomi
requests
urllib3
